refactor: rename source folder to conventional
