+++ src/cmds/fmt-commit.js
